feat(studio): relocate revenue share into Monetisation hub + Team shortcut (#270)
* feat(studio): relocate revenue share into Monetisation hub + Team shortcut

Revenue share is no longer buried as a Settings tab — it's now a tab in the Monetisation hub (Subscriptions · Revenue share). The existing tiers/Connect page becomes the Subscriptions tab; a 308 redirect (preserving ?focus=) covers the old /studio/settings/revenue-share path. Revenue share is now owner-only (Monetisation is an owner-cluster surface; proposing already requires owner).

Team (owner-only) gains a per-member 'Revenue share' status column and a 'Manage revenue share' button to the hub, since agreements are per-creator.

Harness: cap Playwright workers at a flat 2 (was 'CI ? 2 : undefined' = one-per-core locally). The local default exceeded the documented Neon ceiling and made the data-heavy two-actor agreements specs flake on read-after-write under shared-branch contention; verified 3/3 green on a re-seeded DB. Synced e2e/CLAUDE.md.

Closes Codex-4zmw7, Codex-dhxjz. Refs Codex-62gb1, Codex-c066y.


* fix(studio): restore h1.page-title on the Monetisation hub

The relocation moved the page title into the hub +layout.svelte but renamed its class to 'monetisation-hub__title', breaking the subscription E2E specs (01-studio-connect-onboarding, 02-studio-tier-crud) that anchor on 'h1.page-title' after navigating to /studio/monetisation. Restore the 'page-title' class (the studio convention those specs expect). Verified locally: 5 passed (subscription 01+02) on a re-seeded DB.


---------
