+++ apps/web/src/lib/components/studio/dashboard/agreement-focus-items.ts
@@ -222,7 +222,7 @@ export function buildOwnerAgreementFocusItems(
       eyebrow: 'Revenue share',
       title: `${name} sent a counter-proposal`,
       description: `${formatSharePercent(proposal.proposedCreatorSharePercent)} of post-platform ${revLabel} revenue`,
-      href: '/studio/settings/revenue-share',
+      href: '/studio/monetisation/revenue-share',
       tone: 'action',
       icon: EditIcon,
       dismissable: false,
@@ -255,7 +255,7 @@ export function buildOwnerAgreementFocusItems(
       eyebrow: 'Revenue share',
       title: `Set up revenue share with ${name}`,
       description: "They'll get nothing until you propose a split.",
-      href: `/studio/settings/revenue-share?focus=${encodeURIComponent(member.userId)}`,
+      href: `/studio/monetisation/revenue-share?focus=${encodeURIComponent(member.userId)}`,
       tone: 'action',
       icon: EditIcon,
       dismissable: true,
@@ -283,7 +283,7 @@ export function buildOwnerAgreementFocusItems(
       eyebrow: 'Revenue share',
       title: `Agreement with ${name} expires in ${days} day${days === 1 ? '' : 's'}`,
       description: 'Renew or terminate before it auto-expires.',
-      href: `/studio/settings/revenue-share?focus=${encodeURIComponent(agreement.creatorId)}`,
+      href: `/studio/monetisation/revenue-share?focus=${encodeURIComponent(agreement.creatorId)}`,
       tone: 'warning',
       dismissable: true,
     });
